Try Fixing CSRF Token Error

+++ lib/omniauth/strategies/myminifactory.rb
@@ -25,11 +25,15 @@ module OmniAuth
       end
 
       def callback_phase
+        # Verifies CSRF token
+        csrf_token_verified = OmniAuth::RailsCsrfProtection::TokenVerifier.new.call(env)
+        return fail!(:csrf_detected, CallbackError.new(:csrf_detected, 'CSRF detected')) unless csrf_token_verified
+
         # Verifies state token
         state_token_valid = session.delete('omniauth.state') == request.params['state']
         return fail!(:csrf_detected, CallbackError.new(:csrf_detected, 'CSRF detected')) unless state_token_valid
 
-        # Continues with the regular callback phase processing if the state token is valid
+        # Continues with the regular callback phase processing if both tokens are valid
         super
       rescue StandardError => e
         # Handle exceptions
